+++ src/tedx/mpk/data/Stop.java
@@ -4,7 +4,7 @@ import tedx.mpk.Station;
 
 class Stop extends Station {
 	
-	public Stop(String name,long lon,long lat)
+	public Stop(String name,int lon,int lat)
 	{
 		super();
 		setName(name);
